Typescript Conversion, fixes #5 (#6)
This changes the codebase from javascript to typescript. This is partially motivated by wanting to use typescript to help find subtle bugs I might miss, as well as just a good chance to experiment with typescript in general.

It also changes the format of the `destinationLanguages` config option from an array of strings to an array of objects:
```javascript
// Before
"destinationLanguages": ["es", "de"]

// After
"destinationLanguages": [
  { "code": "es" },
  { "code": "de" }
]
```

The config change is motivated by being able to validate it with convict more easily and correctly, as well as opening up for a potential feature where you can specify the cloud translator API by language.